Crate beagle

A basic linear algebra library for computer graphics.

It uses generics to provide extensible square and rectangular Matrix types of sizes 1x1 through 4x4 as well as Vector types of size 1 through 4.

Modules

angle
32-bit and 64-bit type-safe angle units

consts
Traits and structs that emulate integer generics.

index
Indexing and Swizzling for Vec

mat
An row-major array of vectors, written Mat<R, C, T> but pronounced 'matrix'.

num
A collection of numeric traits and functions.

scalar_array
Traits and functions that operate on ScalarArrays.

vec
An array of Scalars, written Vec<D, V> but pronounced 'vector'.

Structs

Value
A simple wrapper to allow generic scalar binary ops

Functions

v
An function to produce Value which is short to type