Module beagle::index

Indexing and Swizzling for Vec

Examples

use beagle::vec::{Vec3,Vec2};
use beagle::index::swizzle::xyzw::*;

let mut v = Vec3::new([8f32, 12f32, 10f32]);
v[YZ] = v[XY] + Vec2::from_value(1f32);
assert_eq!(v, Vec3::new([8f32, 9f32, 13f32]));

Modules

swizzle

Structs

Vec2Ref
This is a temporary structure until something is done with rust-lang/rfcs#997 Do not rely on this type, just rely on the overloaded operators defined for the cases where

Vec3Ref
This is a temporary structure until something is done with rust-lang/rfcs#997 Do not rely on this type, just rely on the overloaded operators defined for the cases where

Vec4Ref
This is a temporary structure until something is done with rust-lang/rfcs#997 Do not rely on this type, just rely on the overloaded operators defined for the cases where

Constants

A
Constant for unary indexing

B
Constant for unary indexing

G
Constant for unary indexing

I0
Constant for unary indexing

I1
Constant for unary indexing

I2
Constant for unary indexing

I3
Constant for unary indexing

P
Constant for unary indexing

Q
Constant for unary indexing

R
Constant for unary indexing

S
Constant for unary indexing

T
Constant for unary indexing

W
Constant for unary indexing

X
Constant for unary indexing

Y
Constant for unary indexing

Z
Constant for unary indexing