Struct beagle::index::Vec4Ref

pub struct Vec4Ref<D, S, I0, I1, I2, I3>(_) where D: Dim<S> + IsLargerThan<I0> + IsLargerThan<I1> + IsLargerThan<I2> + IsLargerThan<I3>, I0: IsSmallerThan<D>, I1: IsSmallerThan<D>, I2: IsSmallerThan<D>, I3: IsSmallerThan<D>, D::Smaller: Array<S>;

This is a temporary structure until something is done with rust-lang/rfcs#997 Do not rely on this type, just rely on the overloaded operators defined for the cases where

Methods

impl<D, S: Clone, I0, I1, I2, I3> Vec4Ref<D, S, I0, I1, I2, I3> where D: Dim<S> + IsLargerThan<I0> + IsLargerThan<I1> + IsLargerThan<I2> + IsLargerThan<I3>, I0: IsSmallerThan<D>, I1: IsSmallerThan<D>, I2: IsSmallerThan<D>, I3: IsSmallerThan<D>, D::Smaller: Array<S>

fn vec(&self) -> Vec<Four, S>