Struct beagle::mat::Mat

pub struct Mat<R, C, S>(_) where C: Dim<S>, R: TwoDim<S, C>, C::Smaller: Array<S>, R::Smaller: Array<C::Type>;

An row-major array of vectors, written Mat<R, C, T> but pronounced 'matrix'.

R represents the number of rows. C represents the number of columns.

Matrices support binary operations between two Matrices or between one Matrix and one Scalar. All Arithmetic operators and Bitwise operators are supported, where the two Scalar types involved support the operation. Most operations operate on each component separately.

The exceptions is matrix multiplied by matrix, which requires performs a linear algebraic multiplication and requires the underlying Scalar types to support Add and Mul.

Matrices can be multiplied by a vector, vector multiplied by matrix, both of which also require the underlying Scalar types to support Add and Mul.

Matrices also support Negation and Logical Negation where the underlying Scalar Type supports it.

Methods

impl<S, C: Dim<S>, R: TwoDim<S, C>> Mat<R, C, S> where C::Smaller: Array<S>, R::Smaller: Array<C::Type>

fn new(m: R::RawType) -> Self where R: Array<C::RawType>, R::RawType: Into<R::Type>, C::RawType: Into<C::Type>

Construct a matrix from an array of arrays

fn from_value(s: S) -> Self where S: Clone, C::Type: Clone

Construct a matrix from a single value

impl<S, C: Dim<S>, R: TwoDim<S, C>> Mat<R, C, S> where C::Smaller: Array<S>, R::Smaller: Array<C::Type>

fn outer_product<Lhs, Rhs>(lhs: Lhs, rhs: Rhs) -> Self where C: Dim<Lhs::Scalar> + Array<Rhs::Scalar>, R: Dim<Rhs::Scalar> + Array<C::Type> + Array<C::Type>, C::Type: Clone, Lhs: VecArrayVal<Row=C>, Rhs: VecArrayVal<Row=R>, Lhs::Scalar: Mul<Rhs::Scalar, Output=S>, Rhs::Scalar: Clone, C::Smaller: Array<Lhs::Scalar>, R::Smaller: Array<Rhs::Scalar>

Constructs a matrix via the outer product of lhs and rhs.