Module beagle::num

A collection of numeric traits and functions.

Structs

AbsoluteEpsilon
An approximator using simple episilon.

Traits

Abs
Types that can have their absolute value taken

Approx
Types that can be used with compared to be approximate equal.

ApproxZero
Types that can be used to compare to zero efficiently.

Clamp
Types that can be clamped

Exp
Types that implement exponential functions

FloatCategory
Types that can be categorized via FpCategory

FloatTransmute
Types that can be transmuted either from a float to an integer representation of the bits, or vice-versa

FractionExponent
Types that can be broken into a fractional component and an exponent

Hyperbolic
Types that implment hyperbolic angle functions.

LoadExponent
Types that can be used to load the exponent of a floating point number

Mix
Types which can be mixed

MulAdd
Types that implement fused multiply-add

Pow
Types that implement the Pow function

Recip
Types that can have the reciprocal taken.

Round
Types that can be rounded.

Sign
Types that can be sign functions.

Sqrt
Types that can be square-rooted.

Step
Types that can generate a step function by comparing two values