Trait beagle::num::FloatCategory

pub trait FloatCategory {
    type Bool;
    type FpCategory;
    fn is_nan(self) -> Self::Bool;
    fn is_infinite(self) -> Self::Bool;
    fn is_finite(self) -> Self::Bool;
    fn is_normal(self) -> Self::Bool;
    fn classify(self) -> Self::FpCategory;
}

Types that can be categorized via FpCategory

Associated Types

type Bool

The bool output type

type FpCategory

The FpCategory output type

Required Methods

fn is_nan(self) -> Self::Bool

Returns true if this value is NaN and false otherwise.

fn is_infinite(self) -> Self::Bool

Returns true if this value is positive infinity or negative infinity and false otherwise.

fn is_finite(self) -> Self::Bool

Returns true if this number is neither infinite nor NaN.

fn is_normal(self) -> Self::Bool

Returns true if the number is neither zero, infinite, [subnormal][subnormal], or NaN.

fn classify(self) -> Self::FpCategory

Returns the floating point category of the number. If only one property is going to be tested, it is generally faster to use the specific predicate instead.