Trait beagle::num::FractionExponent

pub trait FractionExponent {
    type Fraction;
    type Exponent;
    fn frexp(self) -> (Self::Fraction, Self::Exponent);
}

Types that can be broken into a fractional component and an exponent

Associated Types

type Fraction

The floating point fractional component type

type Exponent

The integral exponent component type

Required Methods

fn frexp(self) -> (Self::Fraction, Self::Exponent)

Extracts a floating point significand in the range [0.5, 1.0) and an integer such that: self = Fraction * 2^Exponent