Trait beagle::num::Mix

pub trait Mix<A = Self, Rhs = Self> {
    type Output;
    fn mix(self, y: Rhs, a: A) -> Self::Output;
}

Types which can be mixed

Associated Types

type Output

The output type

Required Methods

fn mix(self, y: Rhs, a: A) -> Self::Output

Performs a linear interpolation between a number and y using a to weight between them. For non-boolean a, the return value is computed as self*(1-a) + y*a*self*(1-a) + y*a. For boolean a, the return value is composed of selected components of x and y. False components of a select the corresponding component of x. True componenets of a select the corresponding components of 'y'.