Trait beagle::num::Round

pub trait Round {
    type Output;
    fn floor(self) -> Self::Output;
    fn ceil(self) -> Self::Output;
    fn trunc(self) -> Self::Output;
    fn fract(self) -> Self::Output;
    fn round(self) -> Self::Output;
}

Types that can be rounded.

Associated Types

type Output

The output type

Required Methods

fn floor(self) -> Self::Output

Returns the largest integer less than or equal to a number.

fn ceil(self) -> Self::Output

Returns the smallest integer greater than or equal to a number.

fn trunc(self) -> Self::Output

Returns the integer part of a number.

fn fract(self) -> Self::Output

Returns the fractional part of a number.

fn round(self) -> Self::Output

Returns the nearest integer to a number. Round half-way cases away from 0.0.