Module beagle::scalar_array

Traits and functions that operate on ScalarArrays.

Modules

array
Helpers for 2d arrays

vec_array
Helpers for 1d arrays

Traits

ConcreteScalarArray
Matrix/Vector types that can be constructed

ConcreteVecArray
Vector types that can be constructed

HasConcreteScalarArray
Matrix/Vector types that have an associated constructable type

HasConcreteVecArray
Vector types that have an associated constructable type

ScalarArray
Types that represent a 2d array of scalars (a matrix or a vector).

ScalarArrayMut
Matrix/Vector types that can be mutated

ScalarArrayRef
Matrix/Vector types that can be references

ScalarArrayVal
Matrix/Vector types that can be deconstructed

VecArrayRef
Vector types that can be references

VecArrayVal
Vector types that can be deconstructed

Functions

apply
Apply a function to all elements of a ScalarArray

apply2
Apply a function to all elements of two ScalarArrays

apply2_mut
Apply a function to all elements of two ScalarArrays

apply2_mut_ref
Apply a function to all elements of two ScalarArrays

apply2_mut_val
Apply a function to all elements of two ScalarArrays

apply2_ref
Apply a function to all elements of two ScalarArrays

apply2_ref_mut
Apply a function to all elements of two ScalarArrays

apply2_ref_val
Apply a function to all elements of two ScalarArrays

apply2_val_mut
Apply a function to all elements of two ScalarArrays

apply2_val_ref
Apply a function to all elements of two ScalarArrays

apply_mut
Apply a function to all elements of a ScalarArray

apply_ref
Apply a function to all elements of a ScalarArray

cpt_cmp
Construct a ScalarArray based on the ordering of the components of two ScalarArray.

cpt_eq
Construct a ScalarArray based on the equality of the components of two ScalarArray.

cpt_ge
Construct a ScalarArray based on the >= comparison of the components of two ScalarArray.

cpt_gt
Construct a ScalarArray based on the > comparison of the components of two ScalarArray.

cpt_le
Construct a ScalarArray based on the <= comparison of the components of two ScalarArray.

cpt_lt
Construct a ScalarArray based on the < comparison of the components of two ScalarArray.

cpt_mul
Multiply two ScalarArrays component-wise.

cpt_mul_assign
Multiply two ScalarArrays component-wise.

cpt_ne
Construct a ScalarArray based on the inequality of the components of two ScalarArray.

cpt_partial_cmp
Construct a ScalarArray based on the partial ordering of the components of two ScalarArray.

fold
Fold all the scalars in a ScalarArray. The first element is mapped with f0, then folding continues with f for other elements.

fold2
Fold all the scalars in two ScalarArrays. The first elements are mapped with f0, then folding continues with f for other elements.

fold2_ref
Fold all the scalars in two ScalarArrays. The first elements are mapped with f0, then folding continues with f for other elements.

fold_mut
Fold all the scalars in a ScalarArray by mutable reference. The first element is mapped with f0, then folding continues with f for other elements.

fold_ref
Fold all the scalars in a ScalarArray by reference. The first element is mapped with f0, then folding continues with f for other elements.

from_scalar
Construct a ScalarArray from a single value

map
Construct a ScalarArray from another ScalarArray and a mapping function

map2
Construct a ScalarArray from two other ScalarArray and a mapping function

map2_mut
Construct a ScalarArray from two other ScalarArray and a mapping function

map2_mut_ref
Construct a ScalarArray from two other ScalarArray and a mapping function

map2_mut_val
Construct a ScalarArray from two other ScalarArray and a mapping function

map2_ref
Construct a ScalarArray from two other ScalarArray and a mapping function

map2_ref_mut
Construct a ScalarArray from two other ScalarArray and a mapping function

map2_ref_val
Construct a ScalarArray from two other ScalarArray and a mapping function

map2_val_mut
Construct a ScalarArray from two other ScalarArray and a mapping function

map2_val_ref
Construct a ScalarArray from two other ScalarArray and a mapping function

map3
Construct a ScalarArray from three other ScalarArray and a mapping function

map_into_2
Construct two ScalarArrays from a ScalarArray and a mapping function

map_mut
Construct a ScalarArray from another ScalarArray by reference and a mapping function

map_ref
Construct a ScalarArray from another ScalarArray by reference and a mapping function

mul_matrix

mul_vector

mul_vector_transpose

sum
Sums all the components of a ScalarArray

transpose
Transposes a ScalarArray