Module beagle::scalar_array::array

Helpers for 2d arrays

Functions

apply
Apply a function to all elements of a 2d array

apply2
Apply a function to all elements of both 2d arrays

fold
Fold all the elements in a 2d array. The first element is mapped with f0, then folding continues with f for other elements.

fold2
Fold all the elements of two 2d arrays. The first elements are mapped with f0, then folding continues with f for other elements.

from_scalar
Construct a 2d array from a single value

map
Construct a 2d array from another 2d array and a mapping function

map2
Construct a 2d array from two other 2d arrays and a mapping function

map3
Construct a 2d array from three other 2d arrays and a mapping function

map_into_2
Construct a 2d array from another 2d array and a mapping function

transpose
Transpose a 2d array