Trait beagle::scalar_array::ScalarArray

pub trait ScalarArray where Self::Row::Smaller: Array<Self::Scalar>, Self::Dim::Smaller: Array<Self::Row::Type> {
    type Scalar;
    type Row: Dim<Self::Scalar>;
    type Dim: TwoDim<Self::Scalar, Self::Row>;
}

Types that represent a 2d array of scalars (a matrix or a vector).

Associated Types

type Scalar

The type of the underlying scalar in the array.

type Row: Dim<Self::Scalar>

The type of a single element of this type (a single row for matrices/a scalar for vectors).

type Dim: TwoDim<Self::Scalar, Self::Row>

The dimension of the scalar array.