Struct beagle::vec::Vec

pub struct Vec<D, S>(_) where D: Dim<S>, D::Smaller: Array<S>;

An array of Scalars, written Vec<D, V> but pronounced 'vector'.

Vectors support binary operations between two Vectors or between one Vector and one Scalar. All Arithmetic operators and Bitwise operators are supported, where the two Scalar types involved support the operation. All operations operate on each component separately.

Vectors also support Negation and Logical Negation where the underlying Scalar Type supports it.

Methods

impl<S, D: Dim<S>> Vec<D, S> where D::Smaller: Array<S>

fn new(v: D::RawType) -> Self where D::RawType: Into<D::Type>

Construct a vector from an array

fn from_value(s: S) -> Self where S: Clone

Construct a Vector from a single value